Save images at end of epoch
Previously it would display images after a specified number of
iterations, but it would name it by epoch. So it would generally save at
some random point in the middle of an epoch, and sometimes twice within
the same epoch, and the second one would overwrite the first. It was
just kind of messy. Now it just saves at the end of each epoch.