Remove a bunch of unused code from dataloaders
The ImageFolder class was totally unused, so I deleted that.
The CustomDatasetDataLoader class was only used in one place, and the
only functionality it added over the base DataLoader was to limit the
number of samples, but that functionality exists separately in the
function that builds the underlying data sets. So I deleted that too,
and moved the little bit of work it did directly into create_dataset.