Don't resize or crop by default
This was setup to resize the image a bit larger, then crop it back down
to size, as part of the training. This was done to increase the variety
of the samples, as every epoch they'd be cropped a bit differently.
However, I don't think that really improves much, and sizing the image
up decreases it's sharpness a little while it's struggling to make the
outputs as sharp as they should be. So, just turn it off by default.