Restructure the UNet module
The UNet module is now much simpler, cleaner and more modular. There are
two new modules, UnetUpBlock and UnetDownBlock, which encompass the set
of things to apply when increasing and decreasing the resolution of the
unet, respectively. The UnetSkipConnectionBlock is now renamed to the
UnetInnerBlock, and it takes a depth rather than flags for innermost or
outermost. The innermost condition is just when the depth is 1. The
outermost is handled in the UnetGenerator class directly.

This makes it much easier to understand and adjust the shape of the
unet.