Unet now uses Upsample and replicate padding
The previous code was producing a subtle checkerboard noise pattern over
all the results. Switching to upsample and a same convolution instead of
a transpose convolution for increasing the resolution addressed that.

It was also producing artifacts along the borders of the output. This
was caused by padding with zeroes, which created the illusion of a sharp
boundary at the edge of the images. Switching to replicate padding mode
means the padding pixels match the neighboring edge pixels, and avoids
these artifacts.