Remove alternate discriminators
The discriminator options were just the pixel discriminator, which is
nearly useless except as a bad baseline for comparison, and the PatchGAN
discriminator, with or without an option for the number of layers. Now
it's just the basic PatchGAN discriminator, without the layer option.
Which means removing the netD commandline option, and the n_layers_D
option as well.