Rearchitect the loss functions
Replace the GANLoss class with BasicGanLoss and WGanLoss classes, which
are much simpler because they don't have to do special cases between the
two every time. The BasicGanLoss takes as an argument which loss
function to use, which will normally be BCEWithLogitsLoss or MSELoss,
but could be others as well. Then added the define_loss function, which
takes the gan_mode from the commandline, and returns the appropriately
initialized loss class, just like the define_G and define_D functions.

Changed the command line arguments to be bce, mse or wgan, instead of
vanilla, lsgan and wgangp, just to be clearer.

Updated the pix2pix model to use this. Also, replaced backwar_D and
backward_G with update_discriminator and update_generator, which differ
from the functions they replace primarily in that they integrate more of
the surrounding code, making the optimize_parameters function simpler.