Output validation set every epoch
This copies over the output format from my main project, which I find
much cleaner. It puts the input, target, prediction and error images all
together into a single image, and outputs the entire validation set for
every epoch.

To make this work, the dataset loading had to be modified a little, to
generation two dataloaders, one for training and one for validation. The
validation loader is passed to the pix2pix_model on initialization, so
it can be accessed for outputing incremental results.